fix(콘텐츠 데이터 관리): 콘텐츠 데이터 삭제
콘텐츠 데이터 검증부분을 일부 수정했습니다. 이 과정에서 TagRepository에 메서드를 추가했으며 테스트코드 또한 변경되었습니다.

+++ src/main/java/com/codeit/playlist/domain/content/service/basic/BasicContentService.java
@@ -121,15 +121,15 @@ public class BasicContentService implements ContentService {
     @Override
     public void delete(UUID contentId) {
         log.debug("컨텐츠 삭제 시작 : id = {}", contentId);
-        Content content = contentRepository.findById(contentId)
-                .orElseThrow(() -> ContentNotFoundException.withId(contentId));
+        if(contentRepository.existsById(contentId)) {
+            log.debug("태그 삭제 시작 : tag = {}", tagRepository.findByContentId(contentId));
+            tagRepository.deleteAllByContentId(contentId); // contentId와 연결된 tags 리스트를 삭제함
+            log.info("태그 삭제 완료 : tag = {}", tagRepository.findByContentId(contentId));
 
-        List<Tag> tags = tagRepository.findByContentId(contentId);
-        log.debug("태그 삭제 시작 : tag = {}", tags);
-        tagRepository.deleteAll(tags); // contentId와 연결된 tags 리스트를 삭제함
-        log.info("태그 삭제 완료 : tag = {}", tags);
-
-        contentRepository.deleteById(contentId);
-        log.info("컨텐츠 삭제 완료 : id = {}", contentId);
+            contentRepository.deleteById(contentId);
+            log.info("컨텐츠 삭제 완료 : id = {}", contentId);
+        } else {
+            throw new ContentNotFoundException();
+        }
     }
 }
